Move decompctx back to the tools folder
